requirements.txt
stripe
PyGithub